Parameterize server/bartender wait-while-idle time
When servers or bartenders have nothing to do (i.e. the seating and
outgoing drink queues are empty in the case of the servers or the
incoming order queue is empty in the case of the bartenders), they'll
sit around doing something else for a fixed period of time before
checking if there's new work to do.

This is a fairly realistic way to model their behavior, since real
servers and bartenders do tend to keep themselves busy when they're
"idle," and there's some overhead between switching from those idle
tasks (checking one's phone, cleaning mugs, sweeping, cleaning tables,
etc.) back to the main tasks.

Previously, the period was just 30 seconds for both, but this commit
adds the `--server-wii` and `--bartender-wwi` options to change these.
(WWI stands for "wait while idle.")

Setting the WWIs to lower values will in general lower the amount of
time servers or bartenders spend in an idle state, BUT there is a
significant effect on performance when the value is set very low (since
new events will be pushed to the event queue more often).